+++ jeap-oauth-mock-server/src/main/java/ch/admin/bit/jeap/oauth/mock/server/config/IntrospectionEndpointAudienceCheck.java
@@ -1,0 +1,5 @@
+package ch.admin.bit.jeap.oauth.mock.server.config;
+
+public enum IntrospectionEndpointAudienceCheck {
+    OFF, ON, WARN
+}